apply requests changes
+ some small animation tweaks
